requirements.txt
torch>=2.0.0
diffusers>=0.28.0
transformers>=4.35.0
accelerate>=0.24.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
safetensors>=0.3.0
huggingface-hub>=0.16.4
datasets>=2.14.0
xformers>=0.0.22
scipy>=1.9.0